+++ sydtest/src/Test/Syd/Path.hs
@@ -20,14 +20,14 @@ tempDirSpec template = setupAround $ tempDirSetupFunc template
 tempDirSetupFunc ::
   -- | Temporary directory name template
   String ->
-  SetupFunc () (Path Abs Dir)
+  SetupFunc (Path Abs Dir)
 tempDirSetupFunc template = makeSimpleSetupFunc $ withSystemTempDir template
 
 tempBinaryFileWithContentsSetupFunc ::
   -- | Temporary directory name template
   String ->
   ByteString ->
-  SetupFunc () (Path Abs File)
+  SetupFunc (Path Abs File)
 tempBinaryFileWithContentsSetupFunc template contents =
   makeSimpleSetupFunc $ \func ->
     withSystemTempFile