+++ sydtest/CHANGELOG.md
@@ -1,11 +1,22 @@
 # Changelog
 
-## [Unreleased]
+## [0.2.0.0] - Unreleased
 
 ### Added
 
 * `beforeWith` and `beforeWith'`
 * `scenarioDir` and `scenarioDirRecur` for scenario testing.
+
+### Changed
+
+* The `SetupFunc` has been simplified to only take one type parameter.
+
+### Deleted
+
+* `composeSetupFunc`, now obsolete: use `<=<` instead.
+* `connectSetupFunc`, now obsolete: use `>=>` instead.
+* `wrapSetupFunc`, now entirely obsolete.
+* `unwrapSetupFunc`, now entirely obsolete.
 
 ## [0.1.0.0] - 2021-03-07